new version
this version will take a binary number of any size and convert it to a
decimal and vice versa. The only restriction is that signed integer was
used so it will only be able to be that large otherwise an unhandled
exception will occur, I may handle this exception in the future.